hook added for multiple use
